refactored screens to have the corresponding names
