+++ app/src/main/java/com/example/buddybudget/Screen.kt
@@ -1,7 +1,7 @@
 package com.example.buddybudget
 
 sealed class Screen(val route: String) {
-    object MainScreen : Screen("main_screen")
+    object HomeScreen : Screen("home_screen")
     object LoginScreen : Screen("login_screen")
     object SplashScreen : Screen("splash_screen")
     object PreLoginScreen :Screen("prelogin_screen")
