+++ app/src/main/java/com/example/buddybudget/SignUpScreen.kt
@@ -1,7 +1,11 @@
 package com.example.buddybudget
 
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
@@ -10,7 +14,11 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -52,5 +60,33 @@ fun SignUpScreen(navController: NavController) {
             label = { Text("Repeat Password") },
             modifier = Modifier.fillMaxWidth()
         )
+
+        Button(
+            //change Screen.MainScreen.route to the create account screen when implemented
+            onClick = {
+                navController.navigate(Screen.SignUpScreen.route)
+            },
+            colors = ButtonDefaults.buttonColors(Color.Cyan),
+            modifier = Modifier
+                .align(Alignment.CenterHorizontally)
+                .fillMaxWidth()
+                .weight(1f, fill = false)
+                .padding(5.dp)
+                .padding(horizontal = 50.dp)
+
+        ) {
+            Text(text = "Create account", fontSize = 25.sp,)
+
+        }
+        
+        LoginButton(navController)
     }
+}
+
+@Composable
+fun LoginButton(navController: NavController) {
+    Text(
+        modifier = Modifier.clickable { navController.navigate(Screen.LoginScreen.route) },
+        text = "Already have an account? Login"
+    )
 }
